requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
PyPDF2
python-docx
pyarrow
textblob
wordcloud
openpyxl
